+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+  * Set SDP port 0 for rejected m-lines #772
   * Handle incoming SSRC 0 BWE probe #769
   * New CryptoProvider API #760
   * Update H.264 profile and level matching logic #758